+++ one/src/events.rs
@@ -1,18 +1,17 @@
-use anyhow::{anyhow, Error, Result};
+use anyhow::{anyhow, Result};
 use ceramic_core::{ssi, Base64UrlString, DidDocument, Jwk};
 use ceramic_p2p::SQLiteBlockStore;
 use ceramic_store::SqlitePool;
 use chrono::{SecondsFormat, TimeZone, Utc};
 use cid::{multibase, multihash, Cid};
 use clap::{Args, Subcommand};
-use enum_as_inner::EnumAsInner;
 use glob::{glob, Paths};
-use minicbor::{data::Tag, data::Type, Decoder};
-use multihash::Multihash;
-use ordered_float::OrderedFloat;
+use crate::CborValue;
 use sqlx::Row;
-use std::{collections::BTreeMap, fs, ops::Index, path::PathBuf, str::FromStr};
+use std::{fs, path::PathBuf, str::FromStr};
 use tracing::debug;
+use crate::ethereum_rpc::eth_transaction_by_hash;
+use multihash::Multihash;
 
 #[derive(Subcommand, Debug)]
 pub enum EventsCommand {
@@ -290,315 +289,7 @@ async fn prev_in_root(
     Ok(prev == current_cid)
 }
 
-async fn eth_transaction_by_hash(cid: Cid, ethereum_rpc_url: &str) -> Result<(Cid, i64)> {
-    // curl https://mainnet.infura.io/v3/{api_token} \
-    //   -X POST \
-    //   -H "Content-Type: application/json" \
-    //   -d '{"jsonrpc":"2.0","method":"eth_getTransactionByHash", "params":["0x{tx_hash}"],"id":1}'
-    let tx_hash = format!("0x{}", hex::encode(cid.hash().digest()));
-    let client = reqwest::Client::new();
-    let res = client
-        .post(ethereum_rpc_url)
-        .json(&serde_json::json!({
-            "jsonrpc": "2.0",
-            "method": "eth_getTransactionByHash",
-            "params": [tx_hash],
-            "id": 1,
-        }))
-        .send()
-        .await?;
-    let json: serde_json::Value = res.json().await?;
-    debug!("txByHash response: {}", json);
-    if let Some(result) = json.get("result") {
-        if let Some(block_hash) = result.get("blockHash") {
-            if let Some(block_hash) = block_hash.as_str() {
-                if let Some(input) = result.get("input") {
-                    if let Some(input) = input.as_str() {
-                        return Ok((
-                            get_root_from_input(input)?,
-                            eth_block_by_hash(block_hash, ethereum_rpc_url).await?,
-                        ));
-                    }
-                }
-            }
-        }
-    }
-    Err(anyhow!("missing fields"))
-}
 
-async fn eth_block_by_hash(block_hash: &str, ethereum_rpc_url: &str) -> Result<i64> {
-    // curl https://mainnet.infura.io/v3/{api_token} \
-    //     -X POST \
-    //     -H "Content-Type: application/json" \
-    //     -d '{"jsonrpc":"2.0","method":"eth_getBlockByHash","params": ["0x{block_hash}",false],"id":1}'
-    let client = reqwest::Client::new();
-    let res = client
-        .post(ethereum_rpc_url)
-        .json(&serde_json::json!({
-            "jsonrpc": "2.0",
-            "method": "eth_getBlockByHash",
-            "params": [block_hash, false],
-            "id": 1,
-        }))
-        .send()
-        .await?;
-    let json: serde_json::Value = res.json().await?;
-    debug!("blockByHash response: {}", json);
-    if let Some(timestamp) = json["result"]["timestamp"].as_str() {
-        get_timestamp_from_hex_string(timestamp)
-    } else {
-        Err(anyhow!("missing fields"))
-    }
-}
-
-fn get_root_from_input(input: &str) -> Result<Cid> {
-    if let Some(input) = input.strip_prefix("0x97ad09eb") {
-        // Strip "0x97ad09eb" from the input and convert it into a cidv1 - dag-cbor - (sha2-256 : 256)
-        // 0x12 -> sha2-256
-        // 0x20 -> 256 bits of hash
-        let root_bytes = [vec![0x12_u8, 0x20], hex::decode(input)?.to_vec()].concat();
-        Ok(Cid::new_v1(0x71, Multihash::from_bytes(&root_bytes)?))
-    } else {
-        Err(anyhow!("input is not anchor-cbor"))
-    }
-}
-
-fn get_timestamp_from_hex_string(ts: &str) -> Result<i64> {
-    // Strip "0x" from the timestamp and convert it to a u64
-    if let Some(ts) = ts.strip_prefix("0x") {
-        Ok(i64::from_str_radix(ts, 16)?)
-    } else {
-        Err(anyhow!("timestamp is not valid hex"))
-    }
-}
-
-#[derive(Eq, PartialEq, Ord, PartialOrd, Clone, Debug, EnumAsInner)]
-pub enum CborValue {
-    // not clear if Indef should be treated distinctly from the finite versions
-    Bool(bool),
-    Null,
-    Undefined,
-    U64(u64),
-    I64(i64),
-    Int(i128), // Cbor ints range from [ -(2^64) , 2^64-1 ] to fit this i65 we use i128
-    F64(OrderedFloat<f64>), // this makes it possible to put CborValue in BTreeMap
-    Simple(u8),
-    Bytes(Vec<u8>),
-    String(String),
-    Array(Vec<CborValue>),
-    Map(BTreeMap<CborValue, CborValue>),
-    Tag((Tag, Box<CborValue>)),
-    Break, // 0xff
-    Unknown(u8),
-}
-
-impl CborValue {
-    fn parse(bytes: &[u8]) -> Result<Self> {
-        let mut decoder = Decoder::new(bytes);
-        let c = CborValue::next(&mut decoder);
-        if decoder.position() != bytes.len() {
-            debug!("decoder not at end {}/{}", decoder.position(), bytes.len());
-            return Err(anyhow!("CborValue decode error not at end"));
-        }
-        c
-    }
-    fn next(decoder: &mut Decoder) -> Result<CborValue> {
-        match decoder.datatype() {
-            Ok(Type::Bool) => Ok(CborValue::Bool(decoder.bool()?)),
-            Ok(Type::Null) => {
-                decoder.null()?;
-                Ok(CborValue::Null)
-            }
-            Ok(Type::Undefined) => {
-                decoder.undefined()?;
-                Ok(CborValue::Undefined)
-            }
-            Ok(Type::U8) => Ok(CborValue::U64(decoder.u8()?.into())),
-            Ok(Type::U16) => Ok(CborValue::U64(decoder.u16()?.into())),
-            Ok(Type::U32) => Ok(CborValue::U64(decoder.u32()?.into())),
-            Ok(Type::U64) => Ok(CborValue::U64(decoder.u64()?)),
-            Ok(Type::I8) => Ok(CborValue::I64(decoder.i8()?.into())),
-            Ok(Type::I16) => Ok(CborValue::I64(decoder.i16()?.into())),
-            Ok(Type::I32) => Ok(CborValue::I64(decoder.i32()?.into())),
-            Ok(Type::I64) => Ok(CborValue::I64(decoder.i64()?)),
-            Ok(Type::Int) => Ok(CborValue::Int(decoder.int()?.into())),
-            Ok(Type::F16) => Ok(CborValue::F64(OrderedFloat(decoder.f16()?.into()))),
-            Ok(Type::F32) => Ok(CborValue::F64(OrderedFloat(decoder.f32()?.into()))),
-            Ok(Type::F64) => Ok(CborValue::F64(OrderedFloat(decoder.f64()?))),
-            Ok(Type::Simple) => Ok(CborValue::Simple(decoder.simple()?)),
-            Ok(Type::Bytes) => Ok(CborValue::Bytes(decoder.bytes()?.to_vec())),
-            Ok(Type::BytesIndef) => Ok(CborValue::Undefined), // TODO: support Type::BytesIndef
-            Ok(Type::String) => Ok(CborValue::String(decoder.str()?.to_string())),
-            Ok(Type::StringIndef) => Ok(CborValue::Undefined), // TODO: support Type::StringIndef
-            Ok(Type::Array) => {
-                let mut array = Vec::new();
-                for _ in 0..decoder.array().unwrap().unwrap() {
-                    let value = CborValue::next(decoder)?;
-                    array.push(value);
-                }
-                Ok(CborValue::Array(array))
-            }
-            Ok(Type::ArrayIndef) => {
-                let mut array = Vec::new();
-                loop {
-                    let value = CborValue::next(decoder)?;
-                    if let CborValue::Break = value {
-                        break;
-                    }
-                    array.push(value)
-                }
-                Ok(CborValue::Array(array))
-            }
-            Ok(Type::Map) => {
-                let mut map = BTreeMap::new();
-                for _ in 0..decoder.map().unwrap().unwrap() {
-                    let key = CborValue::next(decoder)?;
-                    let value = CborValue::next(decoder)?;
-                    map.insert(key, value);
-                }
-                Ok(CborValue::Map(map))
-            }
-            Ok(Type::MapIndef) => Ok(CborValue::Undefined), // TODO: support Type::MapIndef
-            Ok(Type::Tag) => {
-                let tag: Tag = decoder.tag()?;
-                let value = CborValue::next(decoder)?;
-                Ok(CborValue::Tag((tag, Box::new(value))))
-            }
-            Ok(Type::Break) => Ok(CborValue::Break),
-            Ok(Type::Unknown(additional)) => Ok(CborValue::Unknown(additional)),
-            Err(e) => Err(e.into()),
-        }
-    }
-
-    pub fn path(&self, parts: &[&str]) -> CborValue {
-        match parts.split_first() {
-            None => self.clone(), // if there are no parts this is thing at the path.
-            Some((first, rest)) => {
-                let CborValue::Map(map) = &self else {
-                    return CborValue::Undefined; // if self is not a map there is no thing at the path.
-                };
-                let Some(next) = map.get(&first.to_string().into()) else {
-                    return CborValue::Undefined; // if the next part is not in the map there is no thing at the path.
-                };
-                next.path(rest)
-            }
-        }
-    }
-
-    pub fn untag(&self, tag: u64) -> Option<&CborValue> {
-        if let CborValue::Tag((Tag::Unassigned(inner_tag), boxed_value)) = &self {
-            if inner_tag == &tag {
-                return Some(boxed_value);
-            }
-        }
-        None
-    }
-
-    pub fn type_name(&self) -> String {
-        match self {
-            CborValue::Bool(_) => "bool".to_string(),
-            CborValue::Null => "null".to_string(),
-            CborValue::Undefined => "undefined".to_string(),
-            CborValue::U64(_) => "u64".to_string(),
-            CborValue::I64(_) => "i64".to_string(),
-            CborValue::Int(_) => "int".to_string(),
-            CborValue::F64(_) => "f64".to_string(),
-            CborValue::Simple(_) => "simple".to_string(),
-            CborValue::Bytes(_) => "bytes".to_string(),
-            CborValue::String(_) => "string".to_string(),
-            CborValue::Array(_) => "array".to_string(),
-            CborValue::Map(_) => "map".to_string(),
-            CborValue::Tag(_) => "tag".to_string(),
-            CborValue::Break => "break".to_string(),
-            CborValue::Unknown(_) => "unknown".to_string(),
-        }
-    }
-
-    pub fn get_index(&self, index: usize) -> Option<&CborValue> {
-        self.as_array()?.get(index)
-    }
-
-    pub fn get_key(&self, key: &str) -> Option<&CborValue> {
-        self.as_map()?.get(&CborValue::String(key.to_owned()))
-    }
-}
-
-impl From<&[u8]> for CborValue {
-    fn from(bytes: &[u8]) -> Self {
-        CborValue::Bytes(bytes.to_vec())
-    }
-}
-
-impl TryInto<Vec<u8>> for CborValue {
-    type Error = Error;
-    fn try_into(self) -> Result<Vec<u8>, Self::Error> {
-        match self {
-            CborValue::Bytes(bytes) => Ok(bytes.to_vec()),
-            CborValue::String(string) => Ok(string.as_bytes().to_vec()),
-            _ => Err(anyhow!("{} not Bytes or String", self.type_name())),
-        }
-    }
-}
-
-impl From<&str> for CborValue {
-    fn from(string: &str) -> Self {
-        CborValue::String(string.to_string())
-    }
-}
-
-impl From<String> for CborValue {
-    fn from(string: String) -> Self {
-        CborValue::String(string)
-    }
-}
-
-impl TryInto<Cid> for CborValue {
-    type Error = Error;
-    fn try_into(self) -> Result<Cid, Self::Error> {
-        if let Some(CborValue::Bytes(cid_bytes)) = self.untag(42) {
-            Cid::try_from(&cid_bytes[1..]).map_err(|e| e.into())
-        } else {
-            Err(anyhow!("{} not a CID", self.type_name()))
-        }
-    }
-}
-
-impl TryInto<String> for CborValue {
-    type Error = Error;
-    fn try_into(self) -> Result<String, Self::Error> {
-        if let CborValue::String(s) = self {
-            Ok(s)
-        } else {
-            Err(anyhow!("{} not a String", self.type_name()))
-        }
-    }
-}
-
-impl Index<usize> for CborValue {
-    type Output = Self;
-
-    /// Returns a reference to the value corresponding to the supplied key.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the key is not present in the `Vec`.
-    fn index(&self, index: usize) -> &Self {
-        &self.as_array().unwrap()[index]
-    }
-}
-
-impl Index<&CborValue> for CborValue {
-    type Output = Self;
-
-    /// Returns a reference to the value corresponding to the supplied key.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the key is not present in the `BTreeMap`.
-    fn index(&self, key: &CborValue) -> &Self {
-        &self.as_map().unwrap()[key]
-    }
-}
 
 async fn migrate_from_filesystem(input_ipfs_path: PathBuf, store: SQLiteBlockStore) {
     // the block store is split in to 1024 directories and then the blocks stored as files.
@@ -849,7 +540,7 @@ mod tests {
         let ethereum_rpc_url = std::env::var("ETHEREUM_RPC_URL").unwrap();
         assert_eq!(
             validate_time_event(
-                &Cid::try_from("bafyreihu557meceujusxajkaro3epfe6nnzjgbjaxsapgtml7ox5ezb5qy")
+                &Cid::try_from("bafyreihu557meceujusxajkaro3epfe6nnzjgbjaxsapgtml7ox5ezb5qy")  // cspell:disable-line
                     .unwrap(),
                 &block_store,
                 &root_store,
@@ -863,7 +554,7 @@ mod tests {
         // result from the previous call was cached.
         assert_eq!(
             validate_time_event(
-                &Cid::try_from("bafyreihu557meceujusxajkaro3epfe6nnzjgbjaxsapgtml7ox5ezb5qy")
+                &Cid::try_from("bafyreihu557meceujusxajkaro3epfe6nnzjgbjaxsapgtml7ox5ezb5qy") // cspell:disable-line
                     .unwrap(),
                 &block_store,
                 &root_store,
